src/main/java/oliviarla/spring_intro/repository/MemoryMemberRepository.java
package oliviarla.spring_intro.repository;

import oliviarla.spring_intro.domain.Member;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository{

    private static Map<Long, Member> store = new HashMap<>(); //save할 때 저장되는 곳
    private static long sequence = 0L; //key값을 생성해줌
    //동시성 문제를 해결하기 위해 ConcurrentHashMap, accu

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id)); //NULL이어도 optiomal으로 감싸서 반환
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                .filter(member -> member.getName().equals(name))
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore(){
        store.clear();
    }

}
